pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ag-dedicated"
version = "2.0.0"
description = "Analysis of Hawaii's agricultural property tax dedication system across all counties"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["agriculture", "property-tax", "hawaii", "data-analysis", "public-policy"]
authors = [
    {name = "AG-Dedicated Project"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tabula-py>=2.8.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "click>=8.1.0",
    "rich>=13.5.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.9.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
ag-dedicated = "ag_dedicated.cli:main"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | Dedication History
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/ag_dedicated --cov-report=html --cov-report=term"
